// src/packages/types/report/Ser_InvReportCusDebitRpt.ts
export interface Ser_InvReportCusDebitRpt {
  CusName: string;
  GroupAddress: string;
  Phone: string;
  TGD: string;
  PST: string;
  PSG: string;
  TGC: string;
  GroupName: string;
  lst_Ser_InvReportCusDebitRpt: any;
}

export interface Search_Ser_InvReportCusDebitRpt_Params {
  FromDate: string;
  ToDate: string;
  FlagDataWH: any;

  FromDateFromTo: any;
}